✨ role_selection 책임 분리 — role_change.py 추출 (1112→841)
변경 이유:
lifecycle/role_selection.py(1112줄)가 role scoring/selection 과 user-driven
role-change(파싱/적용/audit)라는 두 책임을 한 파일에 담고 있었다(SPLIT_NOW_
PENDING 대상). role-change 서브시스템을 별 모듈로 분리한다.

주요 변경 사항:
- role_change.py 신규(288): RoleChangeRequest/RoleChangeOutcome/
  parse_role_change_request/apply_role_change/append_role_change_audit +
  alias index/verb 상수. scoring 헬퍼(_detect_all_team_request/
  _ensure_tech_lead_first)·상수는 role_selection 에서 단방향 import(순환 없음).
- role_selection.py 에서 해당 블록 제거 → 841줄(1000 미만). format_routing_
  summary(formatting)는 selection 결과 포맷이라 잔류.

비고:
순수 책임 분리 — 동작 불변. scoring 은 role_change 를 import 하지 않아 cycle 없음.
